drf-base64/pyproject.toml
[tool.poetry]
name = "drf-base64"
version = "0.1.0"
description = ""
authors = ["lhy <dev@lhy.kr>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0.4"
django-extensions = "^2.2.9"
drf-yasg = "^1.17.1"
djangorestframework = "^3.11.0"
django-extra-fields = "^2.0.3"
drf-writable-nested = "^0.5.4"
pillow = "^7.0.0"
model_bakery = "^1.1.0"
djangorestframework-camel-case = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
